Correções

+++ src/app/components/header/header.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, TemplateRef, ViewChild } from '@angular/core';
 import { AngularFireAuth } from '@angular/fire/auth';
 import { faUser } from '@fortawesome/free-solid-svg-icons';
 import { AuthService } from 'src/app/services/auth/auth.service';
@@ -8,6 +8,8 @@ import { EMPTY, of } from 'rxjs';
 import { mergeMap, tap } from 'rxjs/operators';
 import { IUser } from 'src/app/interfaces/users';
 import { Router } from '@angular/router';
+import { MatDialog } from '@angular/material/dialog';
+import { ModalComponent, ModalData } from '../modal/modal.component';
 
 @Component({
   selector: 'app-header',
@@ -22,11 +24,20 @@ export class HeaderComponent implements OnInit {
   showMyLinks: boolean;
   user: firebase.User;
 
-  constructor(private afAuth: AngularFireAuth,
-    private fireService: FireService,
-    private authService: AuthService,
-    private route: Router) {
+  @ViewChild('deleteModalHeader') deleteModalHeader: TemplateRef<any>;
 
+  @ViewChild('deleteModalBody') deleteModalBody: TemplateRef<any>;
+
+  constructor(private afAuth: AngularFireAuth, private authService: AuthService, private fireService: FireService, private dialog: MatDialog, private route: Router) {
+    this.afAuth.authState.subscribe(user => {
+      this.user = user;
+
+      if (user && !user.isAnonymous && user.emailVerified)
+        this.isLoggedIn = true;
+      else this.isLoggedIn = false;
+
+      this.showMenu = true;
+    })
   }
 
   ngOnInit(): void {
@@ -64,19 +75,33 @@ export class HeaderComponent implements OnInit {
 
   async excluirConta() {
     const userRef = this.fireService.Firestore.collection("users").doc(this.authService.getCurrentUser.uid);
-
-    const fileRef = this.fireService.Firestore.collection("files").where("userId", "==", this.authService.getCurrentUser.uid);
-
-    fileRef.get().subscribe(fileSnap => {
-      fileSnap.docs.forEach(doc => {
-        doc.ref.delete().subscribe();
-      })
-
+    const dialogRef = this.dialog.open(ModalComponent, {
+      data: {
+        templateHeader: this.deleteModalHeader,
+        templateBody: this.deleteModalBody,
+      } as ModalData
     });
-    await this.authService.SignOut();
-    userRef.delete().subscribe();
-    await this.user.delete();
-    this.goHome();
+
+    dialogRef.componentInstance.userAction.subscribe(
+      async (userConfirm) => {
+        if (userConfirm) {
+          const userRef = this.fireService.Firestore.collection("users").doc(this.authService.getCurrentUser.uid);
+
+          const fileRef = this.fireService.Firestore.collection("files").where("userId", "==", this.authService.getCurrentUser.uid);
+
+          fileRef.get().subscribe(fileSnap => {
+            fileSnap.docs.forEach(doc => {
+              doc.ref.delete().subscribe();
+            })
+
+          });
+          await this.authService.SignOut();
+          userRef.delete().subscribe();
+          await this.user.delete();
+          this.goHome();
+          dialogRef.close();
+        }
+      });
   }
 
   goHome(): void {
